feature/10-create-employee-structure
- Add unit tests for employee services

+++ server/src/Employee/Service/RemoveEmployeeService.php
@@ -2,6 +2,8 @@
 
 namespace Employee\Service;
 
+use Employee\Entity\Employee;
+use Employee\Exception\ResourceNotFoundException;
 use Employee\Repository\EmployeeRepository;
 
 class RemoveEmployeeService
@@ -13,7 +15,9 @@ class RemoveEmployeeService
 
     public function remove(string $email): void
     {
-        $employee = $this->employeeRepository->findOneByEmailOrFail($email);
+        if (null === $employee = $this->employeeRepository->findOneByEmail($email)) {
+            throw ResourceNotFoundException::createFromResourceAndProperty(Employee::class, $email);
+        }
 
         $this->employeeRepository->remove($employee);
     }
